refactor: move compose theme to a dedicated module
