+++ core-compose/build.gradle.kts
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2023 Thibault Seisel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    id("odeon.android.library")
+    id("odeon.android.compose")
+}
+
+android {
+    namespace = "fr.nihilus.music.core.compose"
+}
+
+dependencies {
+    implementation(projects.coreUi)
+    api(libs.compose.material)
+}
